Crate nix_exprc_sys

Structs

BindingsBuilder

EvalState

ExternalValue

ListBuilder

PrimOp

Store

StorePath

__fsid_t

max_align_t

nix_c_context

nix_eval_state_builder

nix_realised_string

nix_value

Enums

ValueType

nix_err

Constants

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

PTRDIFF_MAX

PTRDIFF_MIN

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_LEAST_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_C23_STRTOL

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C23

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C23

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC23

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_TIME_BITS64

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__bool_true_false_are_defined

__glibc_c99_flexarr_available

false_

true_

Functions

nix_alloc_primop⚠

@brief Allocate a PrimOp

nix_alloc_value⚠

@brief Allocate a Nix value

nix_bindings_builder_free⚠

@brief Free a bindings builder

nix_bindings_builder_insert⚠

@brief Insert bindings into a builder @param[out] context Optional, stores error information @param[in] builder BindingsBuilder to insert into @param[in] name attribute name, only used for the duration of the call. @param[in] value value to give the binding @return error code, NIX_OK on success.

nix_c_context_create⚠

nix_c_context_free⚠

nix_clear_err⚠

nix_copy_value⚠

@brief Copy from another value @param[out] context Optional, stores error information @param[out] value Nix value to modify @param[in] source value to copy from @return error code, NIX_OK on success.

nix_err_code⚠

nix_err_info_msg⚠

nix_err_msg⚠

nix_err_name⚠

nix_eval_state_build⚠

nix_eval_state_builder_free⚠

nix_eval_state_builder_load⚠

nix_eval_state_builder_new⚠

nix_eval_state_builder_set_lookup_path⚠

nix_expr_eval_from_string⚠

nix_gc_decref⚠

nix_gc_incref⚠

nix_gc_now⚠

nix_gc_register_finalizer⚠

nix_get_attr_byidx⚠

@brief Get an attribute by index in the sorted bindings

nix_get_attr_byname⚠

@brief Get an attr by name

nix_get_attr_name_byidx⚠

@brief Get an attribute name by index in the sorted bindings

nix_get_attrs_size⚠

@brief Get the element count of an attrset @param[out] context Optional, stores error information @param[in] value Nix value to inspect @return attrset element count, error info via context

nix_get_bool⚠

@brief Get boolean value @param[out] context Optional, stores error information @param[in] value Nix value to inspect @return true or false, error info via context

nix_get_external⚠

@brief Get external reference @param[out] context Optional, stores error information @param[in] value Nix value to inspect @return reference to external, NULL in case of error

nix_get_float⚠

@brief Get float value in 64 bits @param[out] context Optional, stores error information @param[in] value Nix value to inspect @return float contents, error info via context

nix_get_int⚠

@brief Get int value @param[out] context Optional, stores error information @param[in] value Nix value to inspect @return int contents, error info via context

nix_get_list_byidx⚠

@brief Get the ix’th element of a list

nix_get_list_size⚠

@brief Get the length of a list @param[out] context Optional, stores error information @param[in] value Nix value to inspect @return length of list, error info via context

nix_get_path_string⚠

@brief Get path as string @param[out] context Optional, stores error information @param[in] value Nix value to inspect @return string, if the type is NIX_TYPE_PATH @return NULL in case of error.

nix_get_string⚠

@brief Get the raw string

nix_get_type⚠

@addtogroup value_manip Manipulating values @brief Functions to inspect and change Nix language values, represented by nix_value. @{ / /** @anchor getters @name Getters / /@{*/ / @brief Get value type @param[out] context Optional, stores error information @param[in] value Nix value to inspect @return type of nix value

nix_get_typename⚠

@brief Get type name of value as defined in the evaluator @param[out] context Optional, stores error information @param[in] value Nix value to inspect @return type name, owned string @todo way to free the result

nix_has_attr_byname⚠

@brief Check if an attribute name exists on a value @param[out] context Optional, stores error information @param[in] value Nix value to inspect @param[in] state nix evaluator state @param[in] name attribute name @return value, error info via context

nix_init_apply⚠

@brief Set the value to a thunk that will perform a function application when needed.

nix_init_bool⚠

@}*/ /** @name Initializers

nix_init_external⚠

@brief Set an external value @param[out] context Optional, stores error information @param[out] value Nix value to modify @param[in] val the external value to set. Will be GC-referenced by the value. @return error code, NIX_OK on success.

nix_init_float⚠

@brief Set a float @param[out] context Optional, stores error information @param[out] value Nix value to modify @param[in] d the float, 64-bits @return error code, NIX_OK on success.

nix_init_int⚠

@brief Set an int @param[out] context Optional, stores error information @param[out] value Nix value to modify @param[in] i the int @return error code, NIX_OK on success.

nix_init_null⚠

@brief Set null @param[out] context Optional, stores error information @param[out] value Nix value to modify @return error code, NIX_OK on success.

nix_init_path_string⚠

@brief Set a path @param[out] context Optional, stores error information @param[out] value Nix value to modify @param[in] str the path string, copied @return error code, NIX_OK on success.

nix_init_primop⚠

@brief Set primop @param[out] context Optional, stores error information @param[out] value Nix value to modify @param[in] op primop, will be gc-referenced by the value @see nix_alloc_primop @return error code, NIX_OK on success.

nix_init_string⚠

@brief Set a string @param[out] context Optional, stores error information @param[out] value Nix value to modify @param[in] str the string, copied @return error code, NIX_OK on success.

nix_libexpr_init⚠

nix_libstore_init⚠

nix_libstore_init_no_load_config⚠

nix_libutil_init⚠

nix_list_builder_free⚠

@brief Free a list builder

nix_list_builder_insert⚠

@brief Insert bindings into a builder @param[out] context Optional, stores error information @param[in] list_builder ListBuilder to insert into @param[in] index index to manipulate @param[in] value value to insert @return error code, NIX_OK on success.

nix_make_attrs⚠

@brief Create an attribute set from a bindings builder @param[out] context Optional, stores error information @param[out] value Nix value to modify @param[in] b bindings builder to use. Make sure to unref this afterwards. @return error code, NIX_OK on success.

nix_make_bindings_builder⚠

@brief Create a bindings builder @param[out] context Optional, stores error information @param[in] state nix evaluator state @param[in] capacity how many bindings you’ll add. Don’t exceed. @return owned reference to a bindings builder. Make sure to unref when you’re done.

nix_make_list⚠

@brief Create a list from a list builder @param[out] context Optional, stores error information @param[in] list_builder list builder to use. Make sure to unref this afterwards. @param[out] value Nix value to modify @return error code, NIX_OK on success.

nix_make_list_builder⚠

@brief Create a list builder @param[out] context Optional, stores error information @param[in] state nix evaluator state @param[in] capacity how many bindings you’ll add. Don’t exceed. @return owned reference to a list builder. Make sure to unref when you’re done.

nix_realised_string_free⚠

@brief Free a realised string @param[in] realised_string

nix_realised_string_get_buffer_size⚠

@brief Length of the string @param[in] realised_string @return length of the string in bytes

nix_realised_string_get_buffer_start⚠

@brief Start of the string @param[in] realised_string @return pointer to the start of the string. It may not be null-terminated.

nix_realised_string_get_store_path⚠

@brief Get a store path. The store paths are stored in an arbitrary order. @param[in] realised_string @param[in] index index of the store path, must be less than the count @return store path

nix_realised_string_get_store_path_count⚠

@brief Number of realised store paths @param[in] realised_string @return number of realised store paths that were referenced by the string via its context

nix_register_primop⚠

@brief add a primop to the builtins attribute set

nix_set_err_msg⚠

nix_setting_get⚠

nix_setting_set⚠

nix_state_create⚠

nix_state_free⚠

nix_store_copy_closure⚠

nix_store_free⚠

nix_store_get_storedir⚠

nix_store_get_uri⚠

nix_store_get_version⚠

nix_store_is_valid_path⚠

nix_store_open⚠

nix_store_parse_path⚠

nix_store_path_clone⚠

nix_store_path_free⚠

nix_store_path_name⚠

nix_store_real_path⚠

nix_store_realise⚠

nix_string_realise⚠

@brief Realise a string context.

nix_value_call⚠

nix_value_call_multi⚠

nix_value_decref⚠

@brief Decrement the garbage collector reference counter for the given object

nix_value_force⚠

nix_value_force_deep⚠

nix_value_incref⚠

@brief Increment the garbage collector reference counter for the given nix_value.

nix_version_get⚠

Type Aliases

PrimOpFun

@defgroup primops Adding primops @{ / /** @brief Function pointer for primops

Value

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

nix_get_string_callback

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

wchar_t